Auth 2.0

+++ server/resources/review.py
@@ -3,6 +3,8 @@ from http import HTTPStatus
 from flask_restful import Resource, reqparse
 from models import ReviewModel
 
+from .utils import *
+
 review_object_parser = reqparse.RequestParser()
 review_object_parser.add_argument(
     "user_id", type=int, required=True
@@ -35,6 +37,7 @@ class Reviews(Resource):
         return json_response
     
     @classmethod
+    @jwt_required()
     def post(cls, room_id: int):
         request_args = review_object_parser.parse_args()
 
@@ -52,16 +55,26 @@ class ReviewModify(Resource):
     # /reviews/{ review_id }
 
     @classmethod
-    def put(cls, review_id: int):
+    @jwt_required(with_payload=True)
+    def put(cls, review_id: int, payload):
         request_args = review_put_object_parser.parse_args()
         review = ReviewModel.find_by_id(review_id)
+
+        if payload["id"] != review.user_id:
+            return 400, {"access denied"}
+        
         review.review_text = request_args["review_text"]
         review.rate = request_args["rate"]
         return {"message": "Successfully modify review"}, HTTPStatus.ACCEPTED
 
 
     @classmethod
-    def delete(cls, review_id: int):
+    @jwt_required(with_payload=True)
+    def delete(cls, review_id: int, payload):
         review = ReviewModel.find_by_id(review_id)
+
+        if payload["id"] != review.user_id:
+            return 400, {"access denied"}
+        
         review.delete_from_bd()
         return {"message": "Successfully delete review"}, HTTPStatus.OK
